fix clearMediaBloc data and create sliverAppBar for mediaView

+++ lib/src/media/application/mediaFavorites.bloc.dart
@@ -11,6 +11,8 @@ class MediaFavoritesBloc {
   Future<bool> save(MediaModel media) async {
     if (media == null) return false;
 
+    if (await this.isFavorite(media.id.value)) return false;
+
     this._medias.add(media);
     return await this._iMediaRepository.save(media);
   }
@@ -24,6 +26,36 @@ class MediaFavoritesBloc {
   }
 
   Future<bool> deleteById(int id) async {
-    return await this._iMediaRepository.deleteById(id);
+    if (await this._deleteMedia(id))
+      return await this._iMediaRepository.deleteById(id);
+
+    return false;
+  }
+
+  Future<bool> isFavorite(int id) async {
+    bool isIn = false;
+
+    if (this._medias.length == 0) await this.getAll();
+
+    for (MediaModel media in this._medias) {
+      if (media.id.value == id) {
+        isIn = true;
+        break;
+      }
+    }
+
+    return isIn;
+  }
+
+  Future<bool> _deleteMedia(int id) async {
+    if (await isFavorite(id)) {
+      List<MediaModel> aux = [];
+      for (MediaModel media in this._medias) {
+        if (!(media.id.value == id)) aux.add(media);
+      }
+      this._medias = aux;
+      return true;
+    }
+    return false;
   }
 }
